#210 Added scanning support for linux (testing needed)
